services/payment-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
httpx==0.25.2
python-multipart==0.0.6

